Export gettext preprocessor
