+++ i18n-helpers/src/preprocessors.rs
@@ -1,0 +1,2 @@
+mod gettext;
+pub use gettext::Gettext;
